moves action adapters to problem package
